+++ dmr/streaming/sse/metadata.py
@@ -157,6 +157,10 @@ class SSEvent(Generic[_DataT_co]):
             raise ValueError('At least one event field must be non-None')
 
         # Check null byte and new lines:
+        # TODO: this also works when `validate_events` is `False`,
+        # I don't think that it is correct. When `validate_events` is `False`
+        # we need to trust the user to provide valid data and allow everything
+        # for performance reasons.
         check_event_field(id, field_name='id')
         check_event_field(event, field_name='event')