+++ tests/test_unit/test_streaming/test_sse/test_sse_error_handling.py
@@ -5,6 +5,7 @@ from typing import Any
 import pytest
 from typing_extensions import override
 
+from dmr.exceptions import DataRenderingError
 from dmr.plugins.pydantic import PydanticSerializer
 from dmr.streaming import StreamingResponse
 from dmr.streaming.sse import SSEController, SSEvent
@@ -45,3 +46,42 @@ async def test_sse_custom_error_handling(
     assert await get_streaming_content(response) == (
         b'data: 1\r\n\r\nevent: error\r\ndata: zero divizion\r\n\r\n'
     )
+
+
+class _CustomData:
+    """Unserializable."""
+
+
+class _CustomType:
+    """Unserializable."""
+
+    data = _CustomData()
+    event: Any = None
+    id: Any = None
+    comment: Any = None
+    retry: Any = None
+    should_serialize_data = True
+
+
+class _SSECustomType(SSEController[PydanticSerializer]):
+    async def get(self) -> AsyncIterator[Any]:
+        return self._invalid_events()
+
+    async def _invalid_events(self) -> AsyncIterator[Any]:
+        yield _CustomType()
+
+
+@pytest.mark.asyncio
+async def test_sse_event_serialization(
+    dmr_async_rf: DMRAsyncRequestFactory,
+) -> None:
+    """Ensures that we raise on unserializable event types."""
+    request = dmr_async_rf.get('/whatever/')
+
+    response = await dmr_async_rf.wrap(_SSECustomType.as_view()(request))
+
+    assert isinstance(response, StreamingResponse)
+    assert response.streaming
+    assert response.status_code == HTTPStatus.OK
+    with pytest.raises(DataRenderingError, match='_CustomData'):
+        await get_streaming_content(response)